+++ src/containers/AlterForms/index.js
@@ -1,0 +1,3 @@
+export { default as SlideFormNode } from './SlideFormNode';
+export { default as SlideFormEdge } from './SlideFormEdge';
+export { default as SlideFormEgo } from './SlideFormEgo';
