+++ src/containers/Interfaces/AlterForm.js
@@ -12,7 +12,7 @@ import { actionCreators as sessionsActions } from '../../ducks/modules/sessions'
 import { makeNetworkNodesForType } from '../../selectors/interface';
 import { protocolForms } from '../../selectors/protocol';
 import { entityPrimaryKeyProperty } from '../../ducks/modules/network';
-import { SlideFormNode } from '../';
+import { SlideFormNode } from '../AlterForms';
 import defaultMarkdownRenderers from '../../utils/markdownRenderers';
 import { getCSSVariableAsNumber } from '../../ui/utils/CSSVariables';
 
@@ -40,37 +40,36 @@ class AlterForm extends Component {
   getNodeFormName = activeIndex => `NODE_FORM_${activeIndex}`;
 
   formSubmitAllowed = index => (
-    this.swipeRef && this.swipeRef.current.swiper &&
     this.props.formEnabled(this.getNodeFormName(index))
   );
 
   isStageBeginning = () => (
-    this.swipeRef && this.formSubmitAllowed(this.swipeRef.current.swiper.activeIndex) &&
-    this.swipeRef.current.swiper.activeIndex === 0
+    this.state.activeIndex === 0
   );
 
   isStageEnding = () => (
-    this.swipeRef && this.formSubmitAllowed(this.swipeRef.current.swiper.activeIndex) &&
-    this.swipeRef.current.swiper.activeIndex === this.props.stageNodes.length
+    this.formSubmitAllowed(this.state.activeIndex) &&
+    this.state.activeIndex === this.props.stageNodes.length
   );
 
   clickNext = () => {
-    if (this.swipeRef && this.swipeRef.current) {
+    if (this.state.activeIndex > 0) {
       this.props.submitForm(this.getNodeFormName(this.state.activeIndex));
-      if (this.formSubmitAllowed(this.state.activeIndex)) {
-        this.setState({
-          activeIndex: this.state.activeIndex + 1,
-        });
-        this.swipeRef.current.swiper.slideNext();
-      }
+    }
+    if (this.state.activeIndex < this.props.stageNodes.length &&
+      (this.state.activeIndex === 0 || this.formSubmitAllowed(this.state.activeIndex))) {
+      this.setState({
+        activeIndex: this.state.activeIndex + 1,
+      });
+      this.swipeRef.current.swiper.slideNext();
     }
   };
 
   clickPrevious = () => {
-    if (this.swipeRef && this.swipeRef.current) {
-      if (this.formSubmitAllowed(this.state.activeIndex)) {
-        this.props.submitForm(this.getNodeFormName(this.state.activeIndex));
-      }
+    if (this.state.activeIndex > 0 && this.formSubmitAllowed(this.state.activeIndex)) {
+      this.props.submitForm(this.getNodeFormName(this.state.activeIndex));
+    }
+    if (this.state.activeIndex > 0) {
       this.setState({
         activeIndex: this.state.activeIndex - 1,
       });
