+++ src/containers/index.js
@@ -13,6 +13,3 @@ export { default as FormWizard } from './FormWizard';
 export { default as LoadScreen } from './LoadScreen';
 export { default as LoadParamsRoute } from './LoadParamsRoute';
 export { default as CategoricalList } from './CategoricalList';
-export { default as SlideFormNode } from './SlideFormNode';
-export { default as SlideFormEdge } from './SlideFormEdge';
-export { default as SlideFormEgo } from './SlideFormEgo';
